Glue:RunJobs can now be advanced programmatically
