+++ tests/test_glue/test_glue.py
@@ -110,92 +110,6 @@ def test_get_job_exists():
 
 
 @mock_glue
-def test_start_job_run():
-    client = create_glue_client()
-    job_name = create_test_job(client)
-    response = client.start_job_run(JobName=job_name)
-    assert response["JobRunId"]
-
-
-@mock_glue
-def test_start_job_run__multiple_runs_allowed():
-    glue = boto3.client("glue")
-    glue.create_job(
-        Name="somejobname",
-        Role="some-role",
-        ExecutionProperty={"MaxConcurrentRuns": 5},
-        Command={
-            "Name": "some-name",
-            "ScriptLocation": "some-location",
-            "PythonVersion": "some-version",
-        },
-    )
-    for _ in range(5):
-        glue.start_job_run(JobName="somejobname")
-
-    # The 6th should fail
-    with pytest.raises(ClientError) as exc:
-        glue.start_job_run(JobName="somejobname")
-    exc.value.response["Error"]["Code"].should.equal("ConcurrentRunsExceededException")
-    exc.value.response["Error"]["Message"].should.match(
-        "Job with name somejobname already running"
-    )
-
-
-@mock_glue
-def test_start_job_run__single_run_allowed():
-    client = create_glue_client()
-    job_name = create_test_job(client)
-    client.start_job_run(JobName=job_name)
-    with pytest.raises(ClientError) as exc:
-        client.start_job_run(JobName=job_name)
-    exc.value.response["Error"]["Code"].should.equal("ConcurrentRunsExceededException")
-    exc.value.response["Error"]["Message"].should.match(
-        f"Job with name {job_name} already running"
-    )
-
-
-@mock_glue
-def test_get_job_run():
-    client = create_glue_client()
-    job_name = create_test_job(client)
-    job_run_id = client.start_job_run(JobName=job_name)["JobRunId"]
-
-    response = client.get_job_run(JobName=job_name, RunId=job_run_id)
-    response["JobRun"].should.have.key("Id").equals(job_run_id)
-    assert response["JobRun"]["Attempt"]
-    assert response["JobRun"]["PreviousRunId"]
-    assert response["JobRun"]["TriggerName"]
-    assert response["JobRun"]["StartedOn"]
-    assert response["JobRun"]["LastModifiedOn"]
-    assert response["JobRun"]["CompletedOn"]
-    assert response["JobRun"]["JobRunState"]
-    assert response["JobRun"]["Arguments"]
-    assert response["JobRun"]["ErrorMessage"] == ""
-    assert response["JobRun"]["PredecessorRuns"]
-    assert response["JobRun"]["AllocatedCapacity"]
-    assert response["JobRun"]["ExecutionTime"]
-    assert response["JobRun"]["Timeout"]
-    assert response["JobRun"]["MaxCapacity"]
-    assert response["JobRun"]["WorkerType"]
-    assert response["JobRun"]["NumberOfWorkers"]
-    assert response["JobRun"]["SecurityConfiguration"]
-    assert response["JobRun"]["LogGroupName"]
-    assert response["JobRun"]["NotificationProperty"]
-    assert response["JobRun"]["GlueVersion"]
-
-
-@mock_glue
-def test_get_job_run_that_doesnt_exist():
-    client = create_glue_client()
-    job_name = create_test_job(client)
-    with pytest.raises(ClientError) as exc:
-        client.get_job_run(JobName=job_name, RunId="unknown")
-    err = exc.value.response["Error"]
-    err["Code"].should.equal("EntityNotFoundException")
-
-
-@mock_glue
 def test_list_jobs_with_max_results():
     client = create_glue_client()
     create_test_jobs(client, 4)
